raspberry-pi/requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
picamera2==0.3.12
python-dotenv==1.0.0
